pyproject.toml
[tool.poetry]
name = "fluentia"
version = "0.1.0"
description = ""
authors = ["Gabriel Lustosa <lustosaki2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
pydantic = {extras = ["email"], version = "^2.6.2"}
sqlmodel = "^0.0.16"
uvicorn = "^0.27.1"
email-validator = "^2.1.1"
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
taskipy = "^1.12.2"
blue = "^0.9.1"
ruff = "^0.2.2"
httpx = "^0.27.0"
isort = "^5.13.2"
factory-boy = "^3.3.0"
alembic = "^1.13.1"
ipython = "^8.22.1"
pydantic-settings = "^2.2.1"
freezegun = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [".venv", "migrations"]

[tool.ruff.format]
quote-style = "single"

[tool.isort]
profile = "black"
extend_skip = ["migrations"]

[tool.pytest.ini_options]
pythonpath = "."

[tool.blue]
extend-exclude = "(migrations/)"

[tool.taskipy.tasks]
lint = "ruff check . --fix && blue --check . --diff"
format = "ruff format .  && isort ."
run = "uvicorn fluentia.main:app --reload"
pre_test = "task format"
test = "pytest -s -x --cov=. -vv"
post_test = "coverage html"
